refactor: #3 - component 폴더 구조 변경 및 DuoFinderForm 내부 SectionSelector로 컴포넌트 분리
