+++ frontend/ezgg/src/components/layout/Header.jsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import '../styles/Header.css';
+import '../../styles/Header.css';
 
 function Header() {
   return (
