+++ frontend/ezgg/src/components/layout/Layout.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
-import Header from './Header';
-import Footer from './Footer';
-import '../styles/Layout.css';
+import Header from './Header.jsx';
+import Footer from './Footer.jsx';
+import '../../styles/Layout.css';
 
 function Layout() {
   return (
