+++ frontend/ezgg/src/pages/LoginPage.jsx
@@ -1,4 +1,4 @@
-import Login from '../components/layout/Login.jsx';
+import Login from '../components/auth/Login.jsx';
 import '../styles/Login.css';
 
 function LoginPage() {
